+++ java/shared/java/org/signal/libsignal/protocol/incrementalmac/IncrementalMacInputStream.java
@@ -34,18 +34,23 @@ public final class IncrementalMacInputStream extends InputStream {
       byte[] key,
       ChunkSizeChoice sizeChoice,
       byte[] digest,
-      boolean useDirectBuffer) {
+      boolean useDirectBuffer)
+      throws InvalidMacException {
     this.chunkSize = sizeChoice.getSizeInBytes();
     this.currentChunk =
         useDirectBuffer ? ByteBuffer.allocateDirect(chunkSize) : ByteBuffer.allocate(chunkSize);
     this.chunkValidationBuffer =
         this.currentChunk.hasArray() ? null : new byte[VALIDATION_BUFFER_SIZE];
-    this.channel = new MaybeEmptyChannel(channel);
     this.readState = ReadState.READ_FROM_INPUT;
+    this.channel = new MaybeEmptyChannel(channel);
+
+    long handle = Native.ValidatingMac_Initialize(key, chunkSize, digest);
+    if (handle == 0) {
+      throw new InvalidMacException("invalid configuration data");
+    }
 
     this.handleOwner =
-        new NativeHandleGuard.CloseableOwner(
-            Native.ValidatingMac_Initialize(key, chunkSize, digest)) {
+        new NativeHandleGuard.CloseableOwner(handle) {
           @Override
           protected void release(long nativeHandle) {
             Native.ValidatingMac_Destroy(nativeHandle);
@@ -54,7 +59,8 @@ public final class IncrementalMacInputStream extends InputStream {
   }
 
   public IncrementalMacInputStream(
-      InputStream input, byte[] key, ChunkSizeChoice sizeChoice, byte[] digest) {
+      InputStream input, byte[] key, ChunkSizeChoice sizeChoice, byte[] digest)
+      throws InvalidMacException {
     this(Channels.newChannel(input), key, sizeChoice, digest, true);
   }